new updates

+++ src/main/java/com/socgen/employeePortal/controller/EmployeeController.java
@@ -11,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value="/employees")
@@ -20,25 +21,51 @@ public class EmployeeController {
     @Autowired
     EmployeeService employeeService;
 
-    @RequestMapping(method = RequestMethod.GET)
+    @RequestMapping(value="/all",method = RequestMethod.GET)
     @ApiOperation(value="Get All Employees", notes="Gets all Soc Gen Employees", nickname="getEmps")
     public ResponseEntity<List<Employee>> getAllEmployees() {
         List empList = this.employeeService.getAllEmployees();
         return new ResponseEntity<List<Employee>>(empList,new HttpHeaders(),HttpStatus.OK);
     }
 
-    @PostMapping
+    @RequestMapping(value = "/create/{id}",method = RequestMethod.POST)
     @ApiOperation(value="Create Employee", notes="Create Soc Gen Employees", nickname="createEmp")
-    public ResponseEntity<Employee> createOrUpdateEmployee(Employee employee)
+    public ResponseEntity<Employee> createEmployee(@RequestBody  Employee employee)
     {
-        Employee empRec = employeeService.createOrUpdateEmployee(employee);
+        Employee empRec = employeeService.createEmployee(employee);
         return new ResponseEntity<Employee>(empRec, new HttpHeaders(), HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}")
-    @ApiOperation(value="Delete Employee", notes="Delete Soc Gen Employees", nickname="delEmp")
-    public HttpStatus deleteEmployee(@PathVariable("id") Long id)
+    @RequestMapping(value = "/update/{id}",method = RequestMethod.PUT)
+    @ApiOperation(value="Update Employee", notes="Update Soc Gen Employees", nickname="updEmp")
+    public ResponseEntity<Employee> updateEmployee(@RequestBody Employee employee,@PathVariable Long id) throws Exception
     {
+       Optional<Employee> emp =  employeeService.getEmployeeById(id);
+       if (!emp.isPresent())
+            throw new Exception("Cannot find Employee with Id: " + id);
+       System.out.println(employee);
+       if(employee.getFirstName() == null || employee.getFirstName().isEmpty())
+           employee.setFirstName(emp.get().getFirstName());
+       if(employee.getLastName() == null || employee.getLastName().isEmpty())
+           employee.setLastName(emp.get().getLastName());
+       if(employee.getGender() == null || employee.getGender().isEmpty())
+           employee.setGender(emp.get().getGender());
+       if(employee.getDepartment() == null || employee.getDepartment().isEmpty())
+           employee.setDepartment(emp.get().getDepartment());
+       if(employee.getDob() == null )
+           employee.setDob(emp.get().getDob());
+        employee.setId(emp.get().getId());
+        Employee empRec = employeeService.updateEmployee(employee);
+        return new ResponseEntity<Employee>(empRec, new HttpHeaders(), HttpStatus.OK);
+    }
+
+    @DeleteMapping("/delete/{id}")
+    @ApiOperation(value="Delete Employee", notes="Delete Soc Gen Employees", nickname="delEmp")
+    public HttpStatus deleteEmployee(@PathVariable("id") Long id) throws Exception
+    {
+        Optional<Employee> emp =  employeeService.getEmployeeById(id);
+        if (!emp.isPresent())
+            throw new Exception("Cannot find Employee with Id: " + id);
         employeeService.deleteEmployee(id);
         return HttpStatus.FORBIDDEN;
     }
